feat: further progress
